requirements.txt
a2a-sdk[http-server]==0.3.22
langchain==1.2.10
langchain-openai==1.1.8
pydantic==2.11.7
uvicorn==0.34.2
httpx==0.28.1
